requirements.txt
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djoser==2.2.2
